+++ tests/metro/js/metro.js
@@ -1,7 +1,7 @@
 /*
  * Metro 4 Components Library v4.4.0  (https://metroui.org.ua)
  * Copyright 2012-2020 Sergey Pimenov
- * Built at 03/08/2020 12:31:02
+ * Built at 06/08/2020 19:15:02
  * Licensed under MIT
  */
 (function (global, undefined) {
@@ -296,7 +296,7 @@ function hasProp(obj, prop){
         return;
     }
 
-    // 
+    // console.log("Promise polyfill v1.2.0");
 
     var PENDING = 'pending';
     var SEALED = 'sealed';
@@ -603,7 +603,7 @@ function hasProp(obj, prop){
 
 /* global hasProp */
 
-var m4qVersion = "v1.0.7. Built at 25/07/2020 18:43:54";
+var m4qVersion = "v1.0.8. Built at 06/08/2020 18:05:15";
 
 /* eslint-disable-next-line */
 var matches = Element.prototype.matches
@@ -2678,7 +2678,7 @@ $.fn.extend({
                 });
             } else {
                 el.setAttribute(name, val);
-                // 
+                // console.log(name, val);
             }
         });
     },
@@ -3855,14 +3855,14 @@ $.extend({
     hide: function(el, cb){
         var $el = $(el);
 
-        if (el.style.display) {
-            $el.origin('display', (el.style.display ? el.style.display : getComputedStyle(el, null).display));
-        }
+        $el.origin('display', (el.style.display ? el.style.display : getComputedStyle(el, null).display));
         el.style.display = 'none';
+
         if (typeof cb === "function") {
             $.bind(cb, el);
             cb.call(el, arguments);
         }
+
         return this;
     },
 
@@ -4507,7 +4507,7 @@ $.noConflict = function() {
     var Metro = {
 
         version: "4.4.0",
-        compileTime: "03/08/2020 12:31:02",
+        compileTime: "06/08/2020 19:15:02",
         buildNumber: "750",
         isTouchable: isTouch,
         fullScreenEnabled: document.fullscreenEnabled,
@@ -4774,7 +4774,7 @@ $.noConflict = function() {
                         }
 
                     } else  {
-                        //
+                        //console.log(mutation);
                     }
                 });
             };
@@ -5015,6 +5015,7 @@ $.noConflict = function() {
                     this.elem = el;
                     this.element = $(el);
                     this.options = $.extend( {}, defaults, options );
+                    this.component = this.elem;
 
                     this._setOptionsFromDOM();
                     this._runtime();
@@ -5089,6 +5090,14 @@ $.noConflict = function() {
                         element.fire(event.toLowerCase(), data);
 
                     return Utils.exec(o["on"+event], _data, element[0]);
+                },
+
+                getComponent: function(){
+                    return this.component;
+                },
+
+                getComponentName: function(){
+                    return this.name;
                 }
             }, compObj);
 
@@ -19063,6 +19072,193 @@ $.noConflict = function() {
     });
 }(Metro, m4q));
 
+/* eslint-disable */
+(function(Metro, $) {
+    'use strict';
+
+    var LightboxDefaultConfig = {
+        closeIcon: "<span class='default-icon-cross'>",
+        prevIcon: "<span class='default-icon-chevron-left'>",
+        nextIcon: "<span class='default-icon-chevron-right'>",
+        loop: true,
+        source: null,
+        onLightboxCreate: Metro.noop
+    };
+
+    Metro.lightboxSetup = function (options) {
+        LightboxDefaultConfig = $.extend({}, LightboxDefaultConfig, options);
+    };
+
+    if (typeof window["metroLightboxSetup"] !== undefined) {
+        Metro.lightboxSetup(window["metroLightboxSetup"]);
+    }
+
+    Metro.Component('lightbox', {
+        init: function( options, elem ) {
+            this._super(elem, options, LightboxDefaultConfig, {
+                // define instance vars here
+                overlay: null,
+                lightbox: null,
+                current: null,
+                items: []
+            });
+            return this;
+        },
+
+        _create: function(){
+            var that = this, element = this.element, o = this.options;
+
+            this._createStructure();
+            this._createEvents();
+
+            this._fireEvent('lightbox-create');
+        },
+
+        _createStructure: function(){
+            var that = this, element = this.element, o = this.options;
+            var lightbox, overlay;
+
+            overlay = $(".lightbox-overlay");
+
+            if (overlay.length === 0) {
+                overlay = $("<div>").addClass("lightbox-overlay").appendTo("body").hide();
+            }
+
+            lightbox = $("<div>").addClass("lightbox").appendTo("body").hide();
+
+            $("<span>").addClass("lightbox__prev").html(o.prevIcon).appendTo(lightbox);
+            $("<span>").addClass("lightbox__next").html(o.nextIcon).appendTo(lightbox);
+            $("<span>").addClass("lightbox__closer").html(o.closeIcon).appendTo(lightbox);
+            $("<div>").addClass("lightbox__image").appendTo(lightbox);
+
+            this.component = lightbox[0];
+            this.lightbox = lightbox;
+            this.overlay = overlay;
+
+            this._setupItems();
+        },
+
+        _createEvents: function(){
+            var that = this, element = this.element, o = this.options;
+            var lightbox = $(this.component);
+            var items = $(this.items);
+
+            items.on(Metro.events.click, function(){
+                that.open(this);
+            });
+
+            lightbox.on(Metro.events.click, ".lightbox__closer", function(){
+                that.close();
+            });
+
+            lightbox.on(Metro.events.click, ".lightbox__prev", function(){
+                that.prev();
+            });
+
+            lightbox.on(Metro.events.click, ".lightbox__next", function(){
+                that.next();
+            });
+        },
+
+        _setupItems: function(){
+            var that = this, element = this.element, o = this.options;
+            var items = o.source ? $(o.source) : element.children();
+
+            if (items.length === 0) {
+                return ;
+            }
+
+            this.items = items;
+        },
+
+        _goto: function(el){
+            var $el = $(el);
+            var isImage = el.tagName === "IMG";
+            var img = $("<img>"), src;
+            var imageWrapper = this.lightbox.find(".lightbox__image").html("");
+
+            this.current = el;
+
+            if (isImage) {
+                src = $el.attr("data-original") || $el.attr("src");
+                img.attr("src", src);
+                img[0].onload = function(){
+                    var port = this.height > this.width;
+                    img.addClass(port ? "lightbox__image-portrait" : "lightbox__image-landscape");
+                    img.appendTo(imageWrapper);
+                }
+            }
+        },
+
+        _index: function(el){
+            var index = -1;
+
+            this.items.each(function(i){
+                if (this === el) {
+                    index = i;
+                }
+            });
+
+            return index;
+        },
+
+        next: function(){
+            var index, current = this.current;
+
+            index = this._index(current);
+
+            if (index + 1 >= this.items.length) {
+                if (this.options.loop) {
+                    index = -1;
+                } else {
+                    return;
+                }
+            }
+
+            this._goto(this.items[index + 1]);
+        },
+
+        prev: function(){
+            var index, current = this.current;
+
+            index = this._index(current);
+
+            if (index - 1 < 0) {
+                if (this.options.loop) {
+                    index = this.items.length;
+                } else {
+                    return;
+                }
+            }
+
+            this._goto(this.items[index - 1]);
+        },
+
+        open: function(el){
+            var lightbox = $(this.component), overlay = $(this.overlay);
+
+            this._goto(el);
+
+            overlay.show();
+            lightbox.show();
+
+            return this;
+        },
+
+        close: function(){
+            this.overlay.hide();
+            this.lightbox.hide();
+        },
+
+        changeAttribute: function(attr, val){
+        },
+
+        destroy: function(){
+            this.element.remove();
+        }
+    });
+}(Metro, m4q));
+
 (function(Metro, $) {
     'use strict';
     var Utils = Metro.utils;
@@ -26614,7 +26810,7 @@ $.noConflict = function() {
                 })
             }
 
-            this.component = table_component;
+            this.component = table_component[0];
 
             if (o.source !== null) {
 
@@ -29945,7 +30141,7 @@ $.noConflict = function() {
 
                 next = that.slides[that.currentSlide];
 
-                
+                console.log(o.effect.camelCase());
                 if (effects.includes(o.effect)) {
                     Metro.animations[o.effect.camelCase()]($(current), $(next), {duration: o.effectDuration});
                 }